refactor: CourseCard 관련 컴포넌트 파일 구조 변경
