+++ src/entities/course/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './no-liked-course'
