+++ src/entities/course/ui/no-liked-course.tsx
