+++ src/features/course-card/action-bar/api/use-course-like.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useState } from 'react'
 import { useDeleteCourseLike, usePostCourseLike } from '@/src/entities/course'
