+++ src/features/course-card/grid-card.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 import Image from 'next/image'
 import { ProfileImage } from '@/src/shared/ui'
@@ -5,7 +7,7 @@ import { CourseType } from '@/src/entities/course'
 import logo from '@/src/assets/images/(logo)/logo.png'
 import { CourseActionBar, useCourseLike } from '@/src/features'
 import { useEffect, useState } from 'react'
-import { CourseModal } from '../modal/course-modal'
+import { CourseModal } from '@/src/features'
 
 export function CourseGridCard({ course }: { course: CourseType }) {
   const {
