+++ src/features/course-card/index.ts
@@ -1,2 +1,3 @@
 export * from './grid-card'
 export * from './list-card'
+export * from './action-bar'
