--- src/features/course/index.ts
@@ -1,3 +1,0 @@
-export * from './form-course'
-export * from './list-user-course'
-export * from './no-liked-course'
